pyproject.toml
[tool.poetry]
name = "passlib-utils"
version = "0.1.0"
description = """Small extension for Passlib"""

authors = ["Ruslan Ilyasovich Gilfanov <ri.gilfanov@yandex.ru>"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Security",
    "Topic :: Security :: Cryptography",
]

[tool.poetry.dependencies]
python = "^3.8"
argon2-cffi = ">=20.1"
passlib = "^1.7"

[tool.poetry.dev-dependencies]
mypy = "^0.790"
pytest = "^5.2"
peewee = "*"
pony = "*"
sqlalchemy = { git = "https://github.com/sqlalchemy/sqlalchemy.git" }

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
